+++ x/perp/keeper/perp_test.go
@@ -10,10 +10,11 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/stretchr/testify/require"
 
-	nibisimapp "github.com/NibiruChain/nibiru/simapp"
 	"github.com/NibiruChain/nibiru/x/common"
 	"github.com/NibiruChain/nibiru/x/common/asset"
 	"github.com/NibiruChain/nibiru/x/common/testutil"
+	"github.com/NibiruChain/nibiru/x/common/testutil/testapp"
+	"github.com/NibiruChain/nibiru/x/perp/keeper"
 	"github.com/NibiruChain/nibiru/x/perp/types"
 	vpooltypes "github.com/NibiruChain/nibiru/x/vpool/types"
 )
@@ -22,7 +23,7 @@ func TestKeeperClosePosition(t *testing.T) {
 	// TODO(mercilex): simulate funding payments
 	t.Run("success", func(t *testing.T) {
 		t.Log("Setup Nibiru app, pair, and trader")
-		nibiruApp, ctx := nibisimapp.NewTestNibiruAppAndContext(true)
+		nibiruApp, ctx := testapp.NewNibiruTestAppAndContext(true)
 		ctx = ctx.WithBlockTime(time.Now())
 		pair := asset.MustNewPair("xxx:yyy")
 
@@ -44,7 +45,7 @@ func TestKeeperClosePosition(t *testing.T) {
 		require.True(t, vpoolKeeper.ExistsPool(ctx, pair))
 
 		t.Log("Set vpool defined by pair on PerpKeeper")
-		setPairMetadata(nibiruApp.PerpKeeper, ctx, types.PairMetadata{
+		keeper.SetPairMetadata(nibiruApp.PerpKeeper, ctx, types.PairMetadata{
 			Pair:                            pair,
 			LatestCumulativePremiumFraction: sdk.MustNewDecFromStr("0.2"),
 		},
@@ -71,7 +72,7 @@ func TestKeeperClosePosition(t *testing.T) {
 
 		t.Log("open position for bob - long")
 		// force funding payments
-		setPairMetadata(nibiruApp.PerpKeeper, ctx, types.PairMetadata{
+		keeper.SetPairMetadata(nibiruApp.PerpKeeper, ctx, types.PairMetadata{
 			Pair:                            pair,
 			LatestCumulativePremiumFraction: sdk.MustNewDecFromStr("0.3"),
 		})